# Conflicts:
#	src/main/java/com/future/cinemaxx/configuration/DataSetup.java

+++ src/main/java/com/future/cinemaxx/services/genres/GenreServiceImpl.java
@@ -1,5 +1,6 @@
 package com.future.cinemaxx.services.genres;
 
+import com.future.cinemaxx.entities.Category;
 import com.future.cinemaxx.entities.CinemaHall;
 import com.future.cinemaxx.entities.Genre;
 import com.future.cinemaxx.repositories.CinemaHallRepo;
@@ -30,11 +31,22 @@ public class GenreServiceImpl implements GenreServiceInterface{
 
     @Override
     public Genre createGenre(Genre genre) {
+        if(genreRepo.existsByName(genre.getName())){
+            throw new IllegalArgumentException("A category with name: "+ genre.getName()+" already exists");}
         return genreRepo.save(genre);
     }
 
     @Override
+    public Genre updateGenre(int genreId, Genre genre) {
+        Genre updatedGenre = genreRepo.findById(genreId)
+                .orElseThrow(() -> new ResourceNotFoundException());
+        if(genre.getName()!=null){updatedGenre.setName(genre.getName());}
+        return genreRepo.save(updatedGenre);
+    }
+
+    @Override
     public void deleteGenre(int genreId) {
+        if(!genreRepo.existsById(genreId)){throw new ResourceNotFoundException();}
         genreRepo.deleteById(genreId);
     }
 }
